requirements.txt
langchain
langchain-community
streamlit
crewai
python-dotenv
unstructured
email-to